[FAB-4993] Incorrect key usage for issued certs
The primary use for Fabric CA is to issue
enrollment certifcates for use in Fabric.
Enrollment certificates are used to
create digital signatures and as such
the key usage sgould be digital
signature rather than certificate
signing.

This change simply changes the value
in the default config which is generated
when Fabric CA server starts

Change-Id: I219ebeb54e1703961cba2d58ebc3423170a95f69
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>