[FAB-5239] LDAP reconnect for idle timeout
The fabric-ca-server connects to the LDAP server and caches the connection.
If the LDAP server closes the connection because of inactivity, the
fabric-ca-server should reconnect.  That is what this change set does.

The openldap server is reconfigured to close connections after 1 second of
inactivity and the ldap test case to sleep for 3 seconds to hit this
condition.  The test case fails without the code change and passes with
the code change.

Change-Id: I38b06eb987fe939066acf8ebdc4f2d5b81a9b76f
Signed-off-by: Keith Smith <bksmith@us.ibm.com>