[FAB-3581] 1.1: Improve error handling (#1)
This is the first of several change sets which will improve error handling
and logging in the fabric-ca-server.  See [FAB-3581] for a description of
the overall design points.

This change set introduces the "httpErr" object which will be used to
encapsulate errors from the server.  It is defined in servererror.go,
which also contains a list of error codes which will be used in
following change sets.

The servererrors.go file will be deleted in a later change set.

Change-Id: If8fe2e1fd127181ef71270a0f88088f3ffb3b4b5
Signed-off-by: Keith Smith <bksmith@us.ibm.com>