[FAB-3581] Improve error handling (#2)
This is the 2nd phase of improving server error handling.
I introduce two new objects:
1) serverEndpoint - this represents an endpoint.  It performs server-side error
   handling, both for writing the response back to the client and for writing the
   access log entry.
2) serverRequestContext - this represents a single request being processed
   by the server.  The serverEndpoint creates this object when it receives
   a request.  It contains both the http request and response objects
   as well as any state that is associated with the request as it is being
   processed.  For example, it contains the enrollment ID of the caller
   after authentication has been performed.
   Almost all of this code was copied from serverauth.go and so is not new
   code; see the BasicAuthentication and TokenAuthentication functions.

NOTE: I am temporarily decreasing the minimum test coverage from 85% to 80%,
but the next change set will change it back.  The reason for this temporary
lowering of the threshold is because I am introducing new code in this change
set which is not yet called, but in the next change set, existing test cases
will call it.

Change-Id: I763d534083367e6adae642cb8757ef80aebc90ab
Signed-off-by: Keith Smith <bksmith@us.ibm.com>