[FAB-3581] 1.1: Improve error handling (#3)
This change set uses the new serverEndpoint and serverRequestContext
objects which were introduced in previous change sets to handle
the server requests.  Logic that is shared between endpoints is now
in the serverRequestContext object.

The serverauth.go module is deleted because authentication logic was
moved into the serverRequestContext object (in the BasicAuthentication
and TokenAuthentication functions).

The servererrors.go is deleted as all server error logic is now in
servererror.go.

Change-Id: I02a15262bd439a3a3469aea1b1fd336ea8378259
Signed-off-by: Keith Smith <bksmith@us.ibm.com>