[FAB-3013] Reuse connections in the client
Load tests were failing with EOF error from http.Client.
This is because lib.Client contructed transport and
http.Client for every request. By default transport connection
pools are maintained in the transport and are reused. This
was not happening because lib.Client contructed new transport
for each request. This change set fixes this problem by
associating a http.Client object with lib.Client and reusing
it for all requests. This will use one transport for all the
requests there by reusing the connections pooled by the
transport.

Change-Id: I0a12946002ffc653237e0186dbc452684fc56a8a
Signed-off-by: Anil Ambati <aambati@us.ibm.com>