[FAB-5426] Fix unit-tests on vagrant/windows
This change adds error checks on removal of temporary files
to ensure they do get deleted as expected and it adds cleaning
up of the server before to ensure the files are not locked.
It also adds a closing of the DB in case loadCA fails to
avoid leaving the database dangling in the ether.

Change-Id: I6e037cf65df6b61ad556824d72c6e7c7ba64f815
Signed-off-by: Arnaud J Le Hors <lehors@us.ibm.com>