[FAB-5346] Role-based access control (#1)
This is the first change set to add support for RBAC (Role-Based
Access Control) to fabric-ca.  This will be accomplished by supporting
attributes in ECerts (Enrollment Certificates).
This change set introduces the attrmgr component which will be used
in the following change set to put attributes in the ECert.
The APIs support encrypted attributes using BCCSP; however, we
will need a fix to BCCSP to serialize a key before this is really
supported, so the code is commented out for now.

Change-Id: I573d1f65b97768f6f771710d5911747db1e9b73e
Signed-off-by: Keith Smith <bksmith@us.ibm.com>