[FAB-5346] 1.1: Role-based access control (#2)
This change set allows a client to request which attributes should
be added to its ECert when enrolling or reenrolling.
The server processes these attribute requests using the attrmgr
from change set #1 and sends the certificate and secret attribute
info to the client in the response.

The test case sends requests, gets response from the server, and
uses attrmgr to validate that the appropriate attributes are found
and have the correct values.

TODO: Add a test case for reenroll.

Change-Id: I0de57f968f05b29a05be01f0711626d622aed429
Signed-off-by: Keith Smith <bksmith@us.ibm.com>