[FAB-5346] Role-based access control (#3)
Enhance the "fabric-ca-client register" command to specify which attributes
should be added to an ECert by default, if the client requests none specifically
at enroll time.  This allows us to issue ECerts with specific attributes on a
per-user basis without having to modify an application to request them at
enrollment time.

See [FAB-5346] for more info.

Change-Id: Ie6b97deaf7e81e55ec153d1bcd71edadfb6c447b
Signed-off-by: Keith Smith <bksmith@us.ibm.com>