[FAB-5679] Allow empty affiliation string
A user should be allowed to specify an empty string
for affiliation or not be required to specify any
affiliation during registration. In this case,
the user being registered will get the affiliation
of the registrar.

A user can be given root affiliation, assuming the
registrar also has root affiliation, by specifying a
dot (.) in the affiliation request.

All other affiliation requests will continue to enforce
the affiliation hierarchy check on the requested
affiliation with the registrar's affiliation.

Change-Id: Ibcaa593a0e868c1d5ea9a11f6214f33122ead067
Signed-off-by: Saad Karim <skarim@us.ibm.com>