[FAB-5346] Role-based access control (#4)
This change updates the documentation to discuss
Role-Based Access Control (RBAC), or more generally,
Attribute-Based Access Control (ABAC).

I also added a convenience function to the attribute manager
named "True" to check to see if an attribute exists and has a
value of true.  This is useful when using attributes to see
if the user possesses a particular role, as demonstrated by
the test cases and the documentation.

Change-Id: I73456f6132a15299bf7e4e6c10ffe3a0abd4447d
Signed-off-by: Keith Smith <bksmith@us.ibm.com>