[FAB-5786] WIP:DB initialization made more resilient
Improve the database initialization logic to handle
the following scenarios:

1) DB exists but tables have not been created
2) DB exists and table exists but table is empty
3) DB exists, table exists, is non-empty but identity
is not found, then add it

If DB initialization fails the server will  still continue
to start the fabric-ca server. When the next request comes in
from a client, will check to see if DB is initialized,
if not then  will try to initialize the DB before
processing client request.

Change-Id: Ib2ac62d0fa155f6f2b57f59e0d6391fa741490fb
Signed-off-by: Saad Karim <skarim@us.ibm.com>