[FAB-2840] Add home directory configuration
This change will help to improve the usability of
fabric-ca. Originally we had allowed for a custom
configuration file name. This would cause problems if
the configuration file name was not provided for
any subsequent commands after enroll since they won't
know the custom file name that was originally specified.

To prevent confusion the --config option will now be
hidden and replaced with --home, this will still allow
users to specify a configurable home directory. However,
the default configuration file name will be used.

The --config option is still supported but will no longer
be exposed.

See [FAB-2840] for more details.

Change-Id: I0e88d9fa52ae7aa5fbf7d7adfaff6fabd15c6f44
Signed-off-by: Saad Karim <skarim@us.ibm.com>