[ FAB-2919 ] Set postgres SSL environment
function setupSSLClientCertificates() in vendored code
   vendor/github.com/lib/pq/conn.go
calls user.Current() if the environmentvariables
  PGSSLCERT
  PGSSLKEY
are null:
        sslkey := o.Get("sslkey")
        sslcert := o.Get("sslcert")
        if sslkey != "" && sslcert != "" {
           ...
        } else {
                user, err := user.Current()

user.Current in turn, invokes system c code in
   /opt/go/src/os/user/lookup_unix.go
via
   getpwuid_r()

This can cause a sporadic and unpredictable panic due to a gcc bug,
provided that the executable was built with the -static flag (which
is the default).

Consequently an easy workaround for this particular instance is to
merely set the PGSSLCERT and PGSSLKEY environment variables.
(Note that this does not guarantee that other system calls will not
potentially be subject to this malaise.)

This change allows the postgres tests using TLS to be re-enabled.

Change-Id: I3670bb2191ca8591d576ac8bfec2b9d3c4ba6d5e
Signed-off-by: Allen Bailey <eabailey@us.ibm.com>