[FAB-5726] 1. Add client side command to update config
This the first part for adding support for dynamically
updating fabric-ca server's configuration.

This change adds the client side command (cfg) that will be
invoked to request a configuration modifiation on the
server. The 'cfg' command will have 4 flags for adding,
removing, and modifying a server configuration. The fourth
flag will list all the configuation on the server that can
be dynamically modified.

The next change will create the server side handler that will
handle this request.

Change-Id: I26a83b2ebd6e5df99615ae79206db4bdf3d3ca7a
Signed-off-by: Saad Karim <skarim@us.ibm.com>