[FAB-5346] Attribute-based access control (#4)
Enhance the "fabric-ca-client enroll" command to specify which attribute
requests should be sent to the server.  This adds support for the
"--enrollment.attrs" option to the client enroll command.  For example:
   fabric-ca-client enroll --enrollment.attrs "a,b:opt"
overrides any default attributes for the identity.  The ":opt" means that
attribute "b" is optional and will not return an error if the user
doesn't have it.

Change-Id: I9774ef707d4e9df4796c08cf079c7afc8a8e00b7
Signed-off-by: Keith Smith <bksmith@us.ibm.com>