[ FAB-5773 ] Increase ca.go test coverage
As part of identifiying the gaps in the error code coverage,
ca.go UT coverage is increased from 86 to 97+ percent.
The few cases remaining in ca.go are virtually impossible to
create easily and/or naturally. These would include cases
such as DB init/sync errors, BCSP errors and file I/O.

Also, some of the certificates in ./testdata were expired.
Consequently, some of the tests were failing for the wrong
reason. The simplest fix was to regenerate the certificates
and their respective keys.

Change-Id: Id2b3a018e6378ce9e847dfe9529e5af87ff902d9
Signed-off-by: Allen Bailey <eabailey@us.ibm.com>