[FAB-5300] Add support to generate CRL
Added 'fabric-ca-client gencrl' command that provides
the ability to generate Certificate revocation list.

The command accepts optional parameters 'revokedafter',
'revokedbefore', 'expireafter' and 'expirebefore'.
By default, it will generate a CRL that contains all
revoked certificates. The 'revokedafter' and 'revokedbefore'
parameters can be used to generate a CRL that contains
only those certs that were revoked in a period.
The 'expireafter' and 'expirebefore'  parameters can
be used to generate a CRL that contains revoked certs
that expire in a period.

This change set also fixes the problem identified in the
JIRA FAB-6165. When adding test cases and benchmark for
gencrl feature, I encountered this problem. I resolved it
to make progress with this change set by redirecting stderr
to /dev/null in run_benchmarks script, which removed the
need for 'log.SetOutput(ioutil.Discard)' in the TestMain
function.

Change-Id: I033c1ca59b5214a5518e48163191d07307300430
Signed-off-by: Anil Ambati <aambati@us.ibm.com>