[FAB-6332] Save msp dir in default client config file
Currently, the fabric ca client does not store the msp
directory in the default configuration file. When the msp
directory is not the default value (<client home>/msp)
and if -M option is not specified in the subsequent
client commands, it results in an error. This change
set corrects this behavior.

Change-Id: I6891f24f69b71d5bbcabe1f4e1646b282bbb3468
Signed-off-by: Anil Ambati <aambati@us.ibm.com>