[FAB-3416] Map LDAP attrs to fabric CA attrs
This change set adds support for mapping LDAP attributes
to fabric CA attributes.  For example, this provides a way
assign the "hf.Revoker" attribute to LDAP users.

See the documentation in docs/source/users-guide.rst
for a description of how to configure this support.

The FVT ldap_test.sh which is part of CI has been enhanced
to test this.

Change-Id: I0b8a8b0d19d5127a7b5a819a9ae64eee19951690
Signed-off-by: Keith Smith <bksmith@us.ibm.com>