[FAB-6647] 1. Maintain backwards comptability
Need to maintain backwards comptability for any
functionality that is added that affect functionality
of current users.

Version checks are made on the server configuration file,
server executable, and database version. Based on the versions
it is determined if a migration is needed of existing identities
to work with the latest release of fabric-ca-server.

This change set lays down the framework that will check
if migration is need. Next change set will add the
migration logic.

Change-Id: Ie8c2b31aef598d69025664335da96d4a13330cb0
Signed-off-by: Saad Karim <skarim@us.ibm.com>