[FAB-6647] 2. Maintain backwards compatibility
Need to maintain backwards compatibility for any
functionality that is added that affect functionality
of current users.

Version checks are made on the server configuration file,
server executable, and database version. Based on the versions
it is determined if a migration is needed of existing identities
to work with the latest release of fabric-ca-server.

Adds logic that will update existing registrar's in the database
to have the 'hf.Registrar.Attributes' attribute with a value of '*'.

Change-Id: I2f8bd7a610900fc4487966f9eeea7b2b1fc68d4c
Signed-off-by: Saad Karim <skarim@us.ibm.com>