[FAB-5726] 1. Dynamic Cfg - Identities: CLI
This changeset lays down the framework for
the client side command 'identity' that will be used
to either get information on a specific identity,
get all identities that the caller is authorized to
view, or add/modify/remove identities for which the  caller
is authorized.

Next changeset will construct the appropriate HTTP calls
on the client side to send to the server. Also, the
basic framework of the handler on the server side will be
constructed.

Change-Id: If392ced451ab0327cfc656725b2b28b4c939867a
Signed-off-by: Saad Karim <skarim@us.ibm.com>