[FAB-6871] Set OUs in ecerts
This change set removes any OUs requested in the body of an
enrollment request and adds an OU for the identity type
and one OU for each of the parts of the identity's affiliation.
See [FAB-6871] for more of a description.

The test attempts to add a bogus OU and verifies that it is
ignored and we get back the correct OUs.

Change-Id: I1486a39d34cde56d580e63b34fdcab35832ab12d
Signed-off-by: Keith Smith <bksmith@us.ibm.com>