[FAB-5726] 2. Dynamic Cfg - identities: CLI2
This change sets create the appropriate HTTP calls
on the client side for all the various config actions,
such as add, modify, and remove. Also, client side logic
for process the CLI inputs is completed.

Next change set will create the 'identities' handler on
the server side and layout the framework for supporting
adding, modifying, and removing identities on the server.

Change-Id: I0cb5a83dddffebe808aff2c28ed79d0da00f09d9
Signed-off-by: Saad Karim <skarim@us.ibm.com>