[FAB-5726] 6. Dynamic Cfg - Aff: CLI
This changeset lays implements the client side command
'affiliation' that will be used to either get information
on a specific affiliation, get all affiliations that the
caller is authorized to view, or add/modify/remove
affiliations for which the caller is authorized.

Next changeset will construct the handler on the server side
that will be used to handle affiliation related requests.

Change-Id: I0e390623ebe96478dd4221d46d1c8ca0d356275a
Signed-off-by: Saad Karim <skarim@us.ibm.com>