[FAB-6991] Fix max enrollments for bootstrap user
The max enrollments for the bootstrap user should use the
max enrollments setting of the server.  This is needed in order
to allow starting the server to enforce one-time passwords and
also use the default bootstrap user settings.

The additional test case failed prior to this change set,
but now succeeds.

Change-Id: Ibbad7c9e0efd4679138c41277b1f69ebe360132d
Signed-off-by: Keith Smith <bksmith@us.ibm.com>