[FAB-6647] 1. Maintain backwards compatibility
Need to maintain backwards compatibility for any
functionality that is added that affect functionality
of current users.

Version checks are made on the server configuration file,
server executable, and database version. Based on the versions
it is determined if a migration is needed of existing identities
to work with the latest release of fabric-ca-server.

This changeset lays down the framework that will check
if migration is needed. Next changeset will add the
migration logic.

Change-Id: I7ce6bfa421b6e01be9127758bf415448ee6172ce
Signed-off-by: Saad Karim <skarim@us.ibm.com>