[FAB-8494] Make affiliations case insensitive
Affiliations are stored in lowercase in the database.
But when checking if an affiliation exists, server
does case senstive comparison, so it returns an
error even when associating an identity to an
existing affiliation that is in mixed case or uppercase.
This change set fixes this problem. With this change,
an affiliation 'org1' is considered same as 'Org1' or
'ORG1'

Signed-off-by: Anil Ambati <aambati@us.ibm.com>
Change-Id: Ifb1d0eb3c199a6044a0dc993126b367d6ed990a1