[FAB-8585] Added feature to retrieve signing certs
This feature will get the most recent signing certificate for a
given user. The caller needs to be authenticated. The ca-client
was modified to include the new functionality. The signing certs
are pulled via a POST to the CA server or from the ca-client.
There are unit tests for pulling the signing certficate. This
was tested using curl and the ca-client.
The feature is intended to be used to aid in adding new nodes to
a network. Using getcacert and signingcert a MSP can be built on
the new node for all of the organizations needed.
Mark Lesswing (mlesswing@realtors.org) also contributed to this
feature.

Change-Id: I83b9f2c235c0533937c2d31c438ee3b9fe00bbba
Signed-off-by: Robert Miroballi <rmiroballi@gmail.com>