[FAB-7238] 3. CLI Input Validation, Time Parsing, Auth
Validates that the input on the CLI is formatted correctly.
Logic added to determine correct time based on the inputs,
handles duration, RFC3339 format, and "now" as inputs for time.

Logic added that checks to see that the caller is authorized
to request certificates based on the attributes owned by the
caller.

Change-Id: I519d7b079b8c65475e7b4172621fd378437ab035
Signed-off-by: Saad Karim <skarim@us.ibm.com>