[FAB-9392] Refactor client CLI code
Refactor the fabric-ca-client CLI code to support integration tests.
Currently all the CLI related code lives in the "main" package.
However, it is desired that integration test be a self contained package.
To be able to do this requires moving the "RunMain" function outside of
the main package into a package of its own so that it can be referenced
and integration package can call to it.

Change-Id: I80d8f851e45ca558bd36acf944e8066ec4725244
Signed-off-by: Saad Karim <skarim@us.ibm.com>