[FAB-8123] Error out if --cacount is set for int CA
Currently, starting an intemediate CA server with --cacount
fails with this error:

"CN 'fabric-ca-server-ca1' cannot be specified for an intermediate CA"

This is because, server creates number of CAs specified by cacount,
with each CA's CN set to 'fabric-ca-server-ca<num>'. But server also
prohibits specifying CN for intermediate CAs. These two functions
conflict with each other. Moreover, --cacount is only recommended for
development/testing purpose. Instead, --cafiles is recommended for production
environment, which requires the user to provide config file for each
CA hosted on the server. For this reason, we will return an error if --cacount
is specified when Fabric CA server is started as an intermediate CA server and
also document that --cacount option is not applicable to an intermediate CA server

Change-Id: Iea948d00ff82c16dd6af1f89da7e3b4f4fa7e69e
Signed-off-by: Anil Ambati <aambati@us.ibm.com>