[FAB-10474] Changed IsAdmin attr type to bool
IsAdmin is an attribute given to all users requesting
Idemix credential. The response to the idemix/credential
request contains a map of attribute name and value.
Currently, all values are converted to strings and sent as
strings in the map. With this change map is changed to
accept any type of value. So, a boolean value will be stored
in the map for the IsAdmin attribute.

Change-Id: I61b848cdabd65fbae88e8d1ca98f51b77396f154
Signed-off-by: Anil Ambati <aambati@us.ibm.com>