[FABC-459] Optimize GetCertificate request
Created a new function that simplifies the
get certificate request by doing error checking
in a single function. Callers of this function
just have to check for error, rather having to check
to see the correct number of certificate were returned.

Also, improved the modularity of the code by creating
an independent package for the certificate request code.

Change-Id: Ibca3fb78207bb5f441ce6c32ff00467464ee8448
Signed-off-by: Saad Karim <skarim@us.ibm.com>