[FABC-542] More accurate HTTP and error codes
Update HTTP codes where a more precise HTTP code
could be used.

Several places were returning a remote error code of
'0', update these places to return an actual error code.

Introduce a new HTTPErr constructor. This constructor accepts
an error as the first parameter,if error passed in as of type
HTTPErr then the original remote error code is preserved
and only the messages are updated to include any wrapper error
messaging. If error is not type HTTPErr a new HTTPErr is
constructed using the passed in values.

Change-Id: If01c443feb5382b781e71f5f4cab5b1bf0e6ca9f
Signed-off-by: Saad Karim <skarim@us.ibm.com>