[FABC-803] Refactoring DB code
Refactored database code into 3 distinct
packages, each package is a databases the CA supports
(SQLite, MySQL, and PostgresSQL).

There was also a lack of proper unit tests for
most of the database code. Unit tests added for all
the 3 databases.

This restructuring of the database code will be useful
for adding support for metrics.

Change-Id: Idd5ed1c802ac6ba35de76462cbbb0baf0222eceb
Signed-off-by: Saad Karim <skarim@us.ibm.com>