[FABCA-xxxx] Remove dependency on fabric-baseimage
This is the start of removing the Fabric Base Image.

At this time we are choosing to not to port to Alpine
as there is no mechanism other building from source
for installing Postgres 9.6 which greatly extends
the build time of the image.

We will use Ubuntu Xenial as was present in Fabric
Base Image until such a time as move to a new version
of Postgres and port work can be done.

The current build scripts were ported directly from
fabric-baseimage. Future CR's on top of this will prune
unnecessary component installations out of the scripts
to provide a minimal viable product.

Change-Id: I85fdaf185359a5ff1732dafe11ede5c73fb62e51
Signed-off-by: Brett Logan <Brett.T.Logan@ibm.com>