[FABC-863] Test fail with open pipe in temp dir
BCCSP searches for keystores by looping through the specified
directories and reading all files. If there is an open pipe
in the directory, the tests will hang permanently until a
timeout occurs.

Added code to create a subdirectory in the directory where
we create the keystore to ensure a clean directory for
creating the keystore

Signed-off-by: Brett Logan <Brett.T.Logan@ibm.com>
Change-Id: Iac73bb8346c8de1efb67da1bfbea4b6aa1a60738